# infra/gateway/openapi.yaml
swagger: "2.0"
info:
  title: oli-api-gateway
  version: "1.0.0"
schemes:
  - https

# For API Gateway / ESPv2, set host to the managed service name.
host: oli-api-0drj1f1cbrv7k.apigateway.oli-staging-fdbba.cloud.goog
basePath: /
produces:
  - application/json
consumes:
  - application/json
...
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: implicit
    type: oauth2

    # ESPv2 compatibility:
    # Define at least one scope so that route-level `security: []` (public)
    # is honored correctly. Protected operations use `firebase: []` which requires
    # a valid Firebase JWT but does NOT require any specific OAuth scope.
    scopes:
      openid: "ESPv2 compatibility"

    x-google-issuer: "https://securetoken.google.com/oli-staging-fdbba"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "oli-staging-fdbba"
    x-google-jwt-locations:
      - header: "Authorization"
        value_prefix: "Bearer "
      - header: "X-Forwarded-Authorization"
        value_prefix: "Bearer "

# Require Firebase by default (fail-closed)
security:
  - firebase: []
paths:
...
  /users/me/labs/summary:
    options:
      operationId: labsSummaryOptions
      security: []
      responses:
        204:
          description: No Content
    get:
      operationId: labsSummaryGet
      security:
        - firebase: []
      responses:
        200:
          description: Labs category summary with latest biomarker values
        401:
          description: Unauthorized
  /users/me/labs/uploads:
    options:
      operationId: labsUploadsOptions
      security: []
      responses:
        204:
          description: No Content
    get:
      operationId: labsUploadsListGet
      security:
        - firebase: []
      responses:
        200:
          description: Lab PDF uploads list
        401:
          description: Unauthorized
    post:
      operationId: labsUploadsCreatePost
      security:
        - firebase: []
      responses:
        202:
          description: Lab PDF upload accepted
        400:
          description: Invalid body or missing idempotency key
        401:
          description: Unauthorized
  /users/me/labs/uploads/{uploadId}:
    parameters:
      - name: uploadId
        in: path
        required: true
        type: string
    options:
      operationId: labsUploadDetailOptions
      security: []
      responses:
        204:
          description: No Content
    get:
      operationId: labsUploadDetailGet
      security:
        - firebase: []
      responses:
        200:
          description: Lab upload detail with extracted biomarkers
        404:
          description: Not found
        401:
          description: Unauthorized
...